+++ ui/packages/shared/profile/src/ProfileIcicleGraph/IcicleGraph/index.tsx
@@ -15,10 +15,11 @@ import {memo, useEffect, useMemo, useRef, useState} from 'react';
 
 import {scaleLinear} from 'd3-scale';
 
-import {Flamegraph, FlamegraphNode, FlamegraphRootNode} from '@parca/client';
+import {Flamegraph} from '@parca/client';
+import {setHoveringNode, useAppDispatch} from '@parca/store';
 import {selectQueryParam, type NavigateFunction} from '@parca/utilities';
 
-import GraphTooltip, {type HoveringNode} from '../../GraphTooltip';
+import GraphTooltip from '../../GraphTooltip';
 import ColorStackLegend from './ColorStackLegend';
 import {IcicleNode, RowHeight} from './IcicleGraphNodes';
 import useColoredGraph from './useColoredGraph';
@@ -40,9 +41,7 @@ export const IcicleGraph = memo(function IcicleGraph({
   sampleUnit,
   navigateTo,
 }: IcicleGraphProps): JSX.Element {
-  const [hoveringNode, setHoveringNode] = useState<
-    FlamegraphNode | FlamegraphRootNode | undefined
-  >();
+  const dispatch = useAppDispatch();
   const [height, setHeight] = useState(0);
   const svg = useRef(null);
   const ref = useRef<SVGGElement>(null);
@@ -71,12 +70,11 @@ export const IcicleGraph = memo(function IcicleGraph({
   }
 
   return (
-    <div onMouseLeave={() => setHoveringNode(undefined)}>
+    <div onMouseLeave={() => dispatch(setHoveringNode(undefined))}>
       <ColorStackLegend navigateTo={navigateTo} compareMode={compareMode} />
       <GraphTooltip
         unit={sampleUnit}
         total={total}
-        hoveringNode={hoveringNode as HoveringNode}
         contextElement={svg.current}
         strings={coloredGraph.stringTable}
         mappings={coloredGraph.mapping}
@@ -98,7 +96,6 @@ export const IcicleGraph = memo(function IcicleGraph({
               totalWidth={width}
               height={RowHeight}
               setCurPath={setCurPath}
-              setHoveringNode={setHoveringNode}
               curPath={curPath}
               data={coloredGraph.root}
               strings={coloredGraph.stringTable}