+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,104 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.248](https://github.com/parca-dev/parca/compare/@parca/web@0.16.247...@parca/web@0.16.248) (2023-03-16)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.247](https://github.com/parca-dev/parca/compare/@parca/web@0.16.246...@parca/web@0.16.247) (2023-03-16)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.246](https://github.com/parca-dev/parca/compare/@parca/web@0.16.245...@parca/web@0.16.246) (2023-03-16)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.245](https://github.com/parca-dev/parca/compare/@parca/web@0.16.244...@parca/web@0.16.245) (2023-03-16)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.244](https://github.com/parca-dev/parca/compare/@parca/web@0.16.243...@parca/web@0.16.244) (2023-03-15)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.243](https://github.com/parca-dev/parca/compare/@parca/web@0.16.242...@parca/web@0.16.243) (2023-03-14)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.242](https://github.com/parca-dev/parca/compare/@parca/web@0.16.241...@parca/web@0.16.242) (2023-03-14)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.241](https://github.com/parca-dev/parca/compare/@parca/web@0.16.240...@parca/web@0.16.241) (2023-03-14)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.240](https://github.com/parca-dev/parca/compare/@parca/web@0.16.239...@parca/web@0.16.240) (2023-03-14)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.239](https://github.com/parca-dev/parca/compare/@parca/web@0.16.238...@parca/web@0.16.239) (2023-03-13)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.238](https://github.com/parca-dev/parca/compare/@parca/web@0.16.237...@parca/web@0.16.238) (2023-03-13)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.237](https://github.com/parca-dev/parca/compare/@parca/web@0.16.236...@parca/web@0.16.237) (2023-03-07)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.236](https://github.com/parca-dev/parca/compare/@parca/web@0.16.235...@parca/web@0.16.236) (2023-03-07)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.235](https://github.com/parca-dev/parca/compare/@parca/web@0.16.234...@parca/web@0.16.235) (2023-03-07)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.234](https://github.com/parca-dev/parca/compare/@parca/web@0.16.233...@parca/web@0.16.234) (2023-03-07)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.233](https://github.com/parca-dev/parca/compare/@parca/web@0.16.232...@parca/web@0.16.233) (2023-03-07)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.232](https://github.com/parca-dev/parca/compare/@parca/web@0.16.231...@parca/web@0.16.232) (2023-03-06)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.231](https://github.com/parca-dev/parca/compare/@parca/web@0.16.230...@parca/web@0.16.231) (2023-03-06)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.230](https://github.com/parca-dev/parca/compare/@parca/web@0.16.229...@parca/web@0.16.230) (2023-03-04)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.229](https://github.com/parca-dev/parca/compare/@parca/web@0.16.228...@parca/web@0.16.229) (2023-03-04)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.228](https://github.com/parca-dev/parca/compare/@parca/web@0.16.227...@parca/web@0.16.228) (2023-03-04)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.227](https://github.com/parca-dev/parca/compare/@parca/web@0.16.226...@parca/web@0.16.227) (2023-03-04)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.226](https://github.com/parca-dev/parca/compare/@parca/web@0.16.225...@parca/web@0.16.226) (2023-03-02)
+
+**Note:** Version bump only for package @parca/web
+
+## 0.16.225 (2023-03-02)
+
+# 0.16.0 (2023-03-01)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.224](https://github.com/parca-dev/parca/compare/@parca/web@0.16.223...@parca/web@0.16.224) (2023-03-01)
 
 **Note:** Version bump only for package @parca/web