+++ ui/packages/shared/client/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.66 (2023-03-07)
+
+**Note:** Version bump only for package @parca/client
+
+## 0.16.65 (2023-03-02)
+
+# 0.16.0 (2023-03-01)
+
+**Note:** Version bump only for package @parca/client
+
 ## 0.16.64 (2023-02-22)
 
 **Note:** Version bump only for package @parca/client